probably more armel compiler errors
git-svn-id: http://fritzing.googlecode.com/svn/trunk@6069 85f94c43-f63a-0410-9d22-21439860c375

+++ fritzing/src/items/pad.cpp
@@ -89,8 +89,8 @@ QString Pad::makeLayerSvg(ViewLayer::ViewLayerID viewLayerID, double mmW, double
 
 	QString connectAt = m_modelPart->localProp("connect").toString();
 	QRectF terminal;
-	double minW = qMin(1.0, wpx / 3);
-	double minH = qMin(1.0, hpx / 3);
+	double minW = qMin((double) 1.0, wpx / 3);
+	double minH = qMin((double) 1.0, hpx / 3);
 	if (connectAt.compare("center", Qt::CaseInsensitive) == 0) {
 		terminal.setRect(2, 2, wpx, hpx);
 	}
